refactor: support -> common 패키지명 변경
