+++ src/main/java/nextstep/common/AuthenticationPrincipal.java
@@ -1,4 +1,4 @@
-package nextstep.support;
+package nextstep.common;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
