+++ src/main/java/nextstep/config/WebConfiguration.java
@@ -1,7 +1,7 @@
 package nextstep.config;
 
 import lombok.AllArgsConstructor;
-import nextstep.support.AuthenticationPrincipalArgumentResolver;
+import nextstep.common.AuthenticationPrincipalArgumentResolver;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
