+++ src/main/java/nextstep/reservation/ReservationService.java
@@ -4,9 +4,9 @@ import lombok.RequiredArgsConstructor;
 import nextstep.reservation.dto.ReservationRequestDto;
 import nextstep.schedule.Schedule;
 import nextstep.schedule.ScheduleDao;
-import nextstep.support.exception.DuplicateEntityException;
-import nextstep.support.exception.NotExistEntityException;
-import nextstep.support.exception.UnauthorizedException;
+import nextstep.common.exception.DuplicateEntityException;
+import nextstep.common.exception.NotExistEntityException;
+import nextstep.common.exception.UnauthorizedException;
 import nextstep.theme.Theme;
 import nextstep.theme.ThemeDao;
 import org.springframework.stereotype.Service;
