+++ src/test/java/nextstep/reservation/ReservationControllerTest.java
@@ -5,8 +5,8 @@ import io.jsonwebtoken.JwtException;
 import io.restassured.RestAssured;
 import nextstep.auth.JwtTokenProvider;
 import nextstep.reservation.dto.ReservationRequestDto;
-import nextstep.support.exception.NotExistEntityException;
-import nextstep.support.exception.UnauthorizedException;
+import nextstep.common.exception.NotExistEntityException;
+import nextstep.common.exception.UnauthorizedException;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
